refactor: extract shared CLI args into lightweight _cli_args module
Move constants, ARG_HELP, validators, argparse wrappers, and parsers
from frontend_core.py into _cli_args.py (stdlib-only imports). Both
pyrit_shell and pyrit_scan can now reference help text and constants
without the heavy frontend_core import. frontend_core re-exports
everything for backward compatibility.

Also replaces per-method lazy frontend_core imports in PyRITShell
with a single self._fc reference set during background init.
